+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v1.5.2
+
+### `@liveblocks/react`
+
+- Fix return type of `resolveUsers`.
+
+# v1.5.1
+
+- Fixes a bug in the bounds check of the `backgroundKeepAliveTimeout` option.
+
 # v1.5.0
 
 Support multiple RoomProviders, or mixing and matching our React package in the